% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_names.R
\name{comp_names}
\alias{comp_names}
\title{Compare scientific names against NaTron}
\usage{
comp_names(data, conn, scientificName = "scientificName")
}
\arguments{
\item{data}{Dataset containing a column with scientific names.}

\item{conn}{DB connection with access permission (see \code{natron_connect()}).}

\item{scientificName}{The column name which contains the scientific names (defults to 'scientificName')}
}
\value{
Returns a dataframe with your species list alongside the three best guesses for taxon in NaTron
}
\description{
This functions checks the scientificNames column in your dataset against the Taxa register in NaTron and return a wrning when there are species not in the register (either due to misspelling, synonyms, etc.)
}
