% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsert_event.R
\name{f_upsert_event}
\alias{f_upsert_event}
\title{Upserts event data into natron database.}
\usage{
f_upsert_event(conn, event_data)
}
\arguments{
\item{event_data}{Event data to be upserted.}

\item{con}{Database connection object with write permissions.}
}
\value{
Pushes and upserts data to database.
}
\description{
Upserts event data into natron database.
}
\examples{

To follow.


}
