% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsert_m_dataset.R
\name{f_upsert_m_dataset}
\alias{f_upsert_m_dataset}
\title{Upserts metadata into Natron database}
\usage{
f_upsert_m_dataset(conn, m_dataset_data)
}
\arguments{
\item{m_dataset_data}{Metadata to be upserted.}

\item{con}{Database connection object with write permissions}
}
\value{
Pushes and upserts data to database
}
\description{
Upserts metadata into Natron database
}
\examples{

Still to come


}
