% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsert_occurrence.R
\name{f_upsert_occurrence}
\alias{f_upsert_occurrence}
\title{Occurrence upsert}
\usage{
f_upsert_occurrence(conn, occurrence_data)
}
\arguments{
\item{occurrence_data}{Occurrence data to be upserted.}

\item{con}{Database connection object with write permissions.}
}
\value{
Pushes and upserts data to database.
}
\description{
Upsert occurrence into Natron database.
}
\examples{

To come.


}
\seealso{
Other upsert functions: \code{\link{upsert_event}},
  \code{\link{upsert_locations}}
}
\concept{upsert functions}
