% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_new_loc.R
\name{get_new_loc}
\alias{get_new_loc}
\title{Generate new locations}
\usage{
get_new_loc(matched_localities = NA, new_localities,
  matched_localities_toimport = NA)
}
\arguments{
\item{matched_localities}{output from location_check - List of localities with pre-matching locality already in database}

\item{new_localities}{output from location_check - List of new localities}

\item{matched_localities_toimport:}{a vector of the localities which have a match in Natron but you still want to import as new (i.e. new localities, but there is an existing locality within the pre-specified radius)}
}
\value{
3 data frames - 1 with localities that did not need changing, 1 with those that did, 1 with all combined
}
\description{
This next function lets you remove newLocations from the 'possible_matches' table if you think the new locations should be imported into Natron instead of matched with existing ones. The newLocations not removed will not be upserted to NaTron, instead we will get the locationIDs from the altermnative locations and use them in the event table.
}
