% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taxon_ID.R
\name{get_taxonID}
\alias{get_taxonID}
\title{Retrieve taxonIDs from NaTron}
\usage{
get_taxonID(names, conn)
}
\arguments{
\item{names}{A vector of scientific species names. These should already be checked agains the NaTron taxa register to make sure they exist there.}

\item{conn}{A connection object with NaTRON (see \code{?natron_connect})}
}
\value{
Returns a vector of UUIDs corresponding to the scientific species names given in the \emph{names} argument.
}
\description{
This function macthes scientific species names against the NaTron taxa register and returns the corresponding taxonIDs (UUIDs).
}
\examples{
\dontrun{
# get some data:
data("setesdal")

# connect to the database
myConnection <- natron_connect("YOUR-USERNAME-HERE")

# Then get the taxonIDs inserted into the setesdal data:
setesdal$taxonID <- get_taxonID(
                    names = setesdal$scientificName,
                    conn = myConnection)

}
}
