% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_check.R
\name{location_check}
\alias{location_check}
\title{This function is made so one can check if the locations in a dataset ment for publishin in NaTron don't alredy exist in the database. (Often the publisher know if they are pre-existing or not and then he doesn't need this function)}
\usage{
location_check(data, conn, radius)
}
\arguments{
\item{conn}{A connection object with natron (can be easily generated with the natron_connect script)}

\item{radius}{The radius in meters in which to search for preexisting localitites in naTron}

\item{flatt_data}{A flattened, long and standardised dataset that you wish to import into NaTron}
}
\value{
A list with 2 data frames. possible_matches shows all those which have a match within the radius set by the function. no_matches holds those without a match.
}
\description{
This function takes as input a standardised flat and long (as opposed to wide formated) dataset and makes the locations table for it. It then scans the existing  NaTron locations table and returns a list of pre-existing localities that lie within 1000 m radius of the give coordinates. The user then is asked to manully check if not some of these localities can be reused. If so, the user must copy-paste the locationID from the existing NaTron locality
}
