% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_table.R
\name{location_table}
\alias{location_table}
\title{Create locations table}
\usage{
location_table(data, conn)
}
\arguments{
\item{data}{A flattened, long and standardised dataset that you wish to import into NaTron}

\item{conn}{A connection object with NaTRON (see \code{?natron_connect})}
}
\value{
Returns the complete location table as a dataframe consistent with the NaTRON formatting.
}
\description{
This function takes as input a standardised flat and long (as opposed to wide formated) dataset and makes the locations table for it, mirroring the structure of the NaTRON locations table.
}
\examples{
data("setesdal")
myConnection <- natron_connect(myUserName)
myLocationTable <- location_table(setesdal, myConnection)
View(myLocationTable)

}
