% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_locations.R
\name{map_locations}
\alias{map_locations}
\title{Map all locations}
\usage{
map_locations(data, zoom, maptype = "hybrid")
}
\arguments{
\item{data}{New data you wish to plot}

\item{zoom}{Determins how far in you want to zoom on the area. Runs from 1 to 15.}

\item{maptype}{Type of map you want to use. Options available are "terrain", "terrain-background", "satellite", "roadmap", "hybrid" (google maps), "terrain", "watercolor", and "toner" (stamen maps). Defaults to "hybrid".}
}
\value{
Map
}
\description{
\code{map_locations} takes all locations from your data and plots them on a map type of your choice.
}
