% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natronbatchupload.R
\docType{package}
\name{natronbatchupload}
\alias{natronbatchupload}
\alias{natronbatchupload-package}
\title{natronbatchupload}
\description{
a package to facilitate the uploading of datasets to NaTRON.
}
\details{
Please see the vigenette for more details.
}
