% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natronR.R
\docType{package}
\name{natronR}
\alias{natronR}
\alias{natronR-package}
\title{natronR}
\description{
a package to facilitate the uploading of datasets to NaTron.
}
\details{
Please see the vigenette for more details.
}
\examples{
\dontrun{
vignette("user-instructions", package = "natronR")
}
}
