% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radius_scan.R
\name{radius_scan}
\alias{radius_scan}
\title{Find NaTRON locations}
\usage{
radius_scan(locationTable, conn, radius)
}
\arguments{
\item{locationTable}{A dataset containing at least two columns: decimalLongitude and decimalLatitude. Usually you want to put here the loction table returned by \code{location_table()}.}

\item{conn}{A connection object with NaTRON (see \code{?natron_connect}).}

\item{radius}{The radius in meters in which to search for preexisting localitites in NaTron.}
}
\value{
If there are any positive matches the function returns these as a dataframe with 15 columns. The last three columns (starting with 'new') are from the dataset you entered and for which there was found a match in NaTRON. The distance from the entered coordinates and the NaTRON location it is compared to is given in 'distance_km'.
}
\description{
This function uses the decimalLatitude and -Longitude columns from a location table to look for existing entries in the NaTRON locations datable within a given radius.
}
\examples{
\dontrun{
myScan <- radius_scan(myLocationTable, myConnection, 100)
}
}
