% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_and_map_event.R
\name{str_map_events}
\alias{str_map_events}
\title{Structure and map event table}
\usage{
str_map_events(data, conn, location_table)
}
\arguments{
\item{data}{Flat data to be structured.}

\item{conn}{NaTRON connection object with access permission (see \code{?natron_connect}).}

\item{location_table}{location table for flattened data (needs to contain UUIDs as locationIDs).}
}
\value{
Dataframe: event table, mapped to the location table and ready to be upserted to Natron.
}
\description{
Function takes inn a flattended datatable with both terms/colums corresponding exactly to the database columns and datatypes and returns an event dataframe ready to be upserted.
}
\examples{
data("setesdal")
myConnection <- natron_connect("myUserName")
myLocationTable(data = setesdal, conn = myConnection)
#map_location(myLocationTable)
myEvents <- str_map_events(data = setesdal, conn = myConnection, location_table = myLocationTable)
}
