% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_and_map_occurrence.R
\name{str_map_occ}
\alias{str_map_occ}
\title{Structure and map occurence table}
\usage{
str_map_occ(data, conn, location_table)
}
\arguments{
\item{data}{Data to be structured.}

\item{conn}{DB connection with access permission (see \code{natron_connect()}).}

\item{location_table}{location table (see \code{location_table()}).}
}
\value{
Occurrence data (dataframe) that is ready to be upserted to Natron.
}
\description{
Function takes inn a datatable where the columns correspond exactly to the database columns and returns an occurrence dataframe ready to be upserted.
}
\examples{
\dontrun{
data("setesdal")
myConnection <- natron_connect("myUserName")
myLocationTable(data = setesdal, conn = myConnection)
#map_location(myLocationTable)
myOccurences <- str_map_occ(data = setesdal,
                            conn = myConnection,
                            location_table = myLocationTable)
}
}
