% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radius_scan.R
\docType{data}
\name{temp_sql}
\alias{temp_sql}
\title{Find NaTRON lcoations}
\format{An object of class \code{character} of length 1.}
\usage{
temp_sql
}
\arguments{
\item{locationTable}{This is loction table returned by \code{location_table()}}

\item{conn}{A connection object with NaTRON (see \code{?natron_connect})}

\item{radius}{The radius in meters in which to search for preexisting localitites in NaTron}
}
\value{
Returns ...
}
\description{
This function
}
\examples{


}
\keyword{datasets}
