% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsert_event.R
\name{upsert_event}
\alias{upsert_event}
\title{Event data upsert}
\usage{
upsert_event(data, conn)
}
\arguments{
\item{data}{Structured and mapped event table to be upserted.}

\item{con}{Database connection object with write permissions.}
}
\value{
Pushes and upserts data to database. Returns nothings
}
\description{
Upserts event data into Natron database.
}
\examples{

upsert_events(data = myEvents, conn = conn)


}
\seealso{
Other upsert functions: \code{\link{f_upsert_occurrence}},
  \code{\link{upsert_locations}}
}
\concept{upsert functions}
