% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsert_event.R
\name{upsert_events}
\alias{upsert_events}
\title{Event data upsert}
\usage{
upsert_events(data, conn)
}
\arguments{
\item{data}{Structured and mapped event table to be upserted (see \code{?str_map_events()}).}

\item{conn}{Database connection object with write permissions (see \code{?natron_connect}).}
}
\value{
Pushes and upserts data to database. Returns nothing.
}
\description{
Upserts event data to Natron.
}
\examples{
\dontrun{
upsert_events(data = myEvents, conn = myConnection)
}

}
\seealso{
Other upsert functions: \code{\link{upsert_locations}},
  \code{\link{upsert_occ}}
}
\concept{upsert functions}
