% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsert_occurrence.R
\name{upsert_occ}
\alias{upsert_occ}
\title{Occurence data upsert}
\usage{
upsert_occ(data, conn)
}
\arguments{
\item{data}{Structured and mapped occurence table to be upserted (see \code{?str_map_occ()}).}

\item{con}{Database connection object with write permissions (see \code{?natron_connect}).}
}
\value{
Pushes and upserts data to database. Returns nothing.
}
\description{
Upserts occurence data to Natron.
}
\examples{
upsert_occ(data = myOccurences, conn = myConnection)

}
\seealso{
Other upsert functions: \code{\link{upsert_events}},
  \code{\link{upsert_locations}}
}
\concept{upsert functions}
