% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_names.R
\name{comp_names}
\alias{comp_names}
\title{Compare scientific names against NaTron}
\usage{
comp_names(names, conn)
}
\arguments{
\item{names}{A vector of scientific species names.}

\item{conn}{DB connection with access permission (see \code{natron_connect()}).}
}
\value{
Returns a dataframe with your species list alongside the three best gueses for taxon in NaTron
}
\description{
This functions checks the scientificNames column in your dataset against the Taxa register in NaTron and return a warning when there are species not in the register (either due to misspelling, synonyms, etc.)
}
\examples{
\dontrun{
data("setesdal")
myConnection <- natron_connect("YOUR-USERNAME-HERE")

myNames <- comp_names(
    names = setesdal$scientificName,
    conn = myConnection)

View(myNames)
}
}
