% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsert.R
\name{f_upsert_event}
\alias{f_upsert_event}
\title{Upsert events ... must be revritten}
\usage{
f_upsert_event(conn, event_data)
}
\arguments{
\item{event_data}{connetion object with write permissions (conn)}

\item{con}{database connection object}
}
\value{
push and upsert data to database
}
\description{
Upsert events ... must be revritten
}
\examples{

f_upsert_event(conn,event_data)

}
