% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsert_locations.R
\name{f_upsert_location}
\alias{f_upsert_location}
\title{Location upsert}
\usage{
f_upsert_location(conn, location_data)
}
\arguments{
\item{con}{Database connection object with write permissions}

\item{locations}{Location to be upserted.}
}
\value{
Pushes and upserts data to database
}
\description{
Upserts location into Natron database
}
\examples{

Still to come


}
