% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_check.R
\name{location_check}
\alias{location_check}
\title{Location check}
\usage{
location_check(data, conn, radius)
}
\arguments{
\item{data}{A flattened, long and standardised dataset that you wish to import into NaTron}

\item{conn}{A connection object with NaTRON (see \code{?natron_connect})}

\item{radius}{The radius in meters in which to search for preexisting localitites in NaTron}
}
\value{
A list with 2 data frames. *possible_matches* shows all those which have a match within the radius set by the function. *no_matches* holds those without a match.
}
\description{
This function takes as input a standardised flat and long (as opposed to wide formated) dataset and makes the locations table for it. It then scans the existing  NaTron locations table and returns a list of pre-existing localities that lie within a given radius of the coordinates. The user must then manully check if some of these localities can be reused. If so, the user must copy-paste the locationID from the existing NaTron locality
}
