% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_and_map_occurrence.R
\name{str_map_occ}
\alias{str_map_occ}
\title{Function takes inn a flattended datatable with both terms/colums corresponding exactly to the database columns and datatypes and returns an occurrence dataframe ready to be upserted.}
\usage{
str_map_occ(flatt_data, conn, location_table)
}
\arguments{
\item{flatt_data}{Flatenned data to be structured}

\item{conn}{DB connection with access permission, can easily be produced using natron_connect script}

\item{location_table}{location table for flattened data (needs to be run through location_check and get_new_loc).}
}
\value{
Occurrence data that is ready to be upserted to Natron.
}
\description{
Function takes inn a flattended datatable with both terms/colums corresponding exactly to the database columns and datatypes and returns an occurrence dataframe ready to be upserted.
}
